+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-03-22
+
+This release contains breaking changes. See README for more details.
+
+### Added
+
+- New `CspReportControllerBase` controller base class that can be extended
+    - Replaces the old, built-in CSP controller, as the automatic registration was problematic
+    - Example on how to use it
+- `AddCspMediaType`-Method to add the CSP media type to an `IMvcBuilder`
+
+### Removed
+
+- Built-In CSP-Controller, as the automatic registration was problematic
+    - You can use the `CspReportControllerBase` instead, see README
+
 ## [1.3.1] - 2022-02-03
 
 ### Added